+++ app/services/container_services/wordpress_services.rb
@@ -1,1 +1,0 @@
-module ContainerServices::WordpressServices; end