+++ db/migrate/20230704010353_update_network_node_driver.rb
@@ -1,5 +1,6 @@
 class UpdateNetworkNodeDriver < ActiveRecord::Migration[7.0]
   def change
+    return unless Rails.env.production?
     ##
     # We added 'Driver' to the nat rule spec. This will ensure consul has that.
     Node.all.each do |node|