+++ db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema[7.0].define(version: 2023_07_04_010353) do
+ActiveRecord::Schema[7.0].define(version: 2023_09_05_194226) do
   # These are extensions that must be enabled in order to support this database
   enable_extension "plpgsql"
   enable_extension "uuid-ossp"
@@ -390,6 +390,7 @@ ActiveRecord::Schema[7.0].define(version: 2023_07_04_010353) do
     t.boolean "override_autoremove", default: false, null: false
     t.string "category"
     t.bigint "product_id"
+    t.bigint "shm_size", default: 0, null: false
     t.index ["category"], name: "index_container_images_on_category"
     t.index ["container_image_provider_id"], name: "index_container_images_on_container_image_provider_id"
     t.index ["is_load_balancer"], name: "index_container_images_on_is_load_balancer"
@@ -543,6 +544,7 @@ ActiveRecord::Schema[7.0].define(version: 2023_07_04_010353) do
     t.decimal "auto_scale_max", default: "0.0", null: false
     t.boolean "override_autoremove", default: false, null: false
     t.bigint "image_variant_id"
+    t.bigint "shm_size", default: 0, null: false
     t.index ["deployment_id"], name: "index_deployment_container_services_on_deployment_id"
     t.index ["image_variant_id"], name: "index_deployment_container_services_on_image_variant_id"
     t.index ["is_load_balancer"], name: "index_deployment_container_services_on_is_load_balancer"